+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidOptionsInitializer.java
@@ -140,7 +140,8 @@ final class AndroidOptionsInitializer {
 
     // registerActivityLifecycleCallbacks is only available if Context is an AppContext
     if (context instanceof Application) {
-      options.addIntegration(new ActivityLifecycleIntegration((Application) context));
+      options.addIntegration(
+          new ActivityLifecycleIntegration((Application) context, buildInfoProvider));
     } else {
       options
           .getLogger()