+++ sentry-jul/src/main/java/io/sentry/jul/SentryHandler.java
@@ -8,7 +8,6 @@ import io.sentry.SentryLevel;
 import io.sentry.SentryOptions;
 import io.sentry.protocol.Message;
 import io.sentry.protocol.SdkVersion;
-import io.sentry.util.CollectionUtils;
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Date;
@@ -20,6 +19,7 @@ import java.util.logging.Handler;
 import java.util.logging.Level;
 import java.util.logging.LogManager;
 import java.util.logging.LogRecord;
+import java.util.stream.Collectors;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
@@ -185,10 +185,15 @@ public class SentryHandler extends Handler {
     if (throwable != null) {
       event.setThrowable(throwable);
     }
-    final Map<String, String> mdcProperties =
-        CollectionUtils.shallowCopy(MDC.getMDCAdapter().getCopyOfContextMap());
-    if (mdcProperties != null && !mdcProperties.isEmpty()) {
-      event.getContexts().put("MDC", mdcProperties);
+    Map<String, String> mdcProperties = MDC.getMDCAdapter().getCopyOfContextMap();
+    if (mdcProperties != null) {
+      mdcProperties =
+          mdcProperties.entrySet().stream()
+              .filter(it -> it.getValue() != null)
+              .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+      if (!mdcProperties.isEmpty()) {
+        event.getContexts().put("MDC", mdcProperties);
+      }
     }
     event.setExtra(THREAD_ID, record.getThreadID());
     return event;