+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/SentrySpanRestTemplateCustomizerTest.kt
@@ -8,6 +8,7 @@ import io.sentry.SentryOptions
 import io.sentry.SentryTracer
 import io.sentry.SpanStatus
 import io.sentry.TransactionContext
+import java.io.IOException
 import kotlin.test.Test
 import org.assertj.core.api.Assertions.assertThat
 import org.springframework.http.HttpMethod
@@ -20,21 +21,26 @@ import org.springframework.web.client.RestTemplate
 
 class SentrySpanRestTemplateCustomizerTest {
     class Fixture {
+        val sentryOptions = SentryOptions()
         val hub = mock<IHub>()
         val restTemplate = RestTemplate()
         var mockServer = MockRestServiceServer.createServer(restTemplate)
         val transaction = SentryTracer(TransactionContext("aTransaction", "op", true), hub)
         internal val customizer = SentrySpanRestTemplateCustomizer(hub)
 
-        fun getSut(isTransactionActive: Boolean, status: HttpStatus = HttpStatus.OK): RestTemplate {
+        init {
+            whenever(hub.options).thenReturn(sentryOptions)
+        }
+
+        fun getSut(isTransactionActive: Boolean, status: HttpStatus = HttpStatus.OK, throwIOException: Boolean = false): RestTemplate {
             customizer.customize(restTemplate)
 
             if (isTransactionActive) {
-                val scope = Scope(SentryOptions())
+                val scope = Scope(sentryOptions)
                 scope.setTransaction(transaction)
                 whenever(hub.span).thenReturn(transaction)
 
-                mockServer.expect(MockRestRequestMatchers.requestTo("/test/123"))
+                val scenario = mockServer.expect(MockRestRequestMatchers.requestTo("/test/123"))
                     .andExpect(MockRestRequestMatchers.method(HttpMethod.GET))
                     .andExpect {
                         // must have trace id from the parent transaction and must not contain spanId from the parent transaction
@@ -42,7 +48,13 @@ class SentrySpanRestTemplateCustomizerTest {
                             .endsWith("-1")
                             .doesNotContain(transaction.spanContext.spanId.toString())
                     }
-                    .andRespond(MockRestResponseCreators.withStatus(status).body("OK").contentType(MediaType.APPLICATION_JSON))
+                if (throwIOException) {
+                    scenario.andRespond {
+                        throw IOException()
+                    }
+                } else {
+                    scenario.andRespond(MockRestResponseCreators.withStatus(status).body("OK").contentType(MediaType.APPLICATION_JSON))
+                }
             } else {
                 mockServer.expect(MockRestRequestMatchers.requestTo("/test/123"))
                     .andExpect(MockRestRequestMatchers.method(HttpMethod.GET))
@@ -72,7 +84,22 @@ class SentrySpanRestTemplateCustomizerTest {
     fun `when transaction is active and response code is not 2xx, creates span with error status around RestTemplate HTTP call`() {
         try {
             fixture.getSut(isTransactionActive = true, status = HttpStatus.INTERNAL_SERVER_ERROR).getForObject("/test/{id}", String::class.java, 123)
-        } catch (e: Throwable) {}
+        } catch (e: Throwable) {
+        }
+        assertThat(fixture.transaction.spans).hasSize(1)
+        val span = fixture.transaction.spans.first()
+        assertThat(span.operation).isEqualTo("http.client")
+        assertThat(span.description).isEqualTo("GET /test/123")
+        assertThat(span.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
+        fixture.mockServer.verify()
+    }
+
+    @Test
+    fun `when transaction is active and throws IO exception, creates span with error status around RestTemplate HTTP call`() {
+        try {
+            fixture.getSut(isTransactionActive = true, throwIOException = true).getForObject("/test/{id}", String::class.java, 123)
+        } catch (e: Throwable) {
+        }
         assertThat(fixture.transaction.spans).hasSize(1)
         val span = fixture.transaction.spans.first()
         assertThat(span.operation).isEqualTo("http.client")